Purged render.cpp
